完成有赞

+++ IService/IRoleService.cs
@@ -4,10 +4,10 @@ namespace IService
 {
     public interface IRoleService:IServiceSupport
     {
-        long AddNew(string roleName);
-        void Update(long roleId, string roleName);
-        void MarkDeleted(long id);
-        RoleDTO GetById(long id);
+        int AddNew(string roleName);
+        void Update(int roleId, string roleName);
+        void MarkDeleted(int id);
+        RoleDTO GetById(int id);
         RoleDTO GetByName(string name);
         RoleDTO[] GetAll();
 
@@ -16,10 +16,10 @@ namespace IService
         /// </summary>
         /// <param name="adminUserId"></param>
         /// <param name="roleIds"></param>
-        void AddRoles(long adminUserId, long[] roleIds);
+        void AddRoles(int adminUserId, int[] roleIds);
         //更新权限,先删除再添加
-        void UpdateRoleIds(long adminUserId, long[] roleIds);
+        void UpdateRoleIds(int adminUserId, int[] roleIds);
         //获取用户角色
-        RoleDTO[] GetByAdminUserId(long adminUserId);
+        RoleDTO[] GetByAdminUserId(int adminUserId);
     }
 }
